Introduce noop OVN Router scheduler
OVN implemented Ha Chassis Groups which is recommended way [0]
to configure HA for LRP and uses built in mechanism to provide
HA capabilities.

This patch introduce ``noop`` sheduler which does nothing
from scheduling standpint. The ``ha_chassis_group`` is
configured for all LRPs in this case.
Also ``ha_chassis_group`` sheduling helps to fix various
bugs related to Baremetal and SRIOV ports.

This is first patch in a serie and the end goal is to remove
scheduling based on ``gateway_chasssis` completely.
Which will be implemented in next patches.

Co-Authored-By: <Anton Arefiev> aarefiev@mirantis.com
Co-Authored-By: <Vasyl Saienko> vsaienko@mirantis.com

[0] https://man7.org/linux/man-pages/man5/ovn-nb.5.html

Related-Prod: PRODX-50467

Related-Bug: #2092271
Related-Bug: #2052468
Related-Bug: #1875852

Change-Id: I718276e94fd0b6f32090a7778a8c3643da7779e8
Signed-off-by: Michal Nasiadka <mnasiadka@gmail.com>

Conflicts:
	neutron/common/ovn/constants.py
	neutron/conf/plugins/ml2/drivers/ovn/ovn_conf.py
	neutron/plugins/ml2/drivers/ovn/mech_driver/ovsdb/maintenance.py
	neutron/plugins/ml2/drivers/ovn/mech_driver/ovsdb/ovn_client.py
	neutron/scheduler/l3_ovn_scheduler.py
	neutron/tests/unit/plugins/ml2/drivers/ovn/mech_driver/ovsdb/test_maintenance.py