WIP == Maintenance method to update the LRP from GC to HCG
The new maintenance method
``migrate_lrp_gateway_chassis_to_ha_chassis_group``, replaces the
``Logical_Router_Port`` ``Gateway_Chassis`` with the equivalent
``HA_Chassis_Group``, that is the preferred method for HA scheduling in
OVN, as reported in the LP bug.

NOTE: this patch is still in WIP. This patch cannot be merged separately
from the patch (or patches) chaging the OVN L3 schedulers, that change
the HA method to ``HA_Chassis_Group``, same as this patch.

Partial-Bug: #2092271
Change-Id: I8994415f97bca575b7104824ef889da0235c9a8b
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>