Use HA_Chassis_Group in the OVN L3 scheduler
The NB API method ``schedule_new_gateway`` now creates a
``HA_Chassis_Group`` and its ``HA_Chassis`` registers, and associates it
to the gateway ``Logical_Router_Port``. The usage og ``Gateway_Chassis``
is deprecated.

The option ``redirect-chassis`` is no longer used nor supported in
Neutron. This option was superseded in [1] (v20.03.0) in favor of
``HA_Chassis_Group`` and it was deprecated in [2] (v20.12.0).

This patch also implementes a new maintenance method
``migrate_lrp_gateway_chassis_to_ha_chassis_group``, that replaces the
``Logical_Router_Port`` ``Gateway_Chassis`` with the equivalent
``HA_Chassis_Group``, that is the preferred method for HA scheduling in
OVN, as reported in the LP bug.

[1]https://github.com/ovn-org/ovn/commit/8b7972dbc6777f45d0c5a1039708836d107f534a
[2]https://github.com/ovn-org/ovn/commit/f77db41dfdf900442139c111fe267eb9af696102

Partial-Bug: #2092271
Change-Id: If38f405428a8b27bf551db48c4e96d75deeb09fc
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>

Use HA_Chassis_Group in the OVN L3 scheduler

This patch adds the last bits of this feature:
* Updates the documentation.
* Adds a release note.

With this patch, the feature described in LP#2092271 can be
considered as finished.

Closes-Bug: #2092271

Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I1283bb140bb8242f9053830dac8ec9437e00e695