[OVN] Always exclude the local chassis creating ``HA_Chassis_Group``
When a ``HA_Chassis_Group`` for an external port is created, the method
always discards the chassis bound to the port. A logical port can have
several chassis stored in the ``Port_Binding`` register, in the ``name``
field (the main one) and in the ``additional_chassis`` list (see method
``get_chassis_host_for_port``).

Closes-Bug: #2125569
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I3346c2a16f69c3b16d1693db42335b133a721bf9