[OVN] Add method ``sync_ha_chassis_group_network_unified``
This new method is meant to be used when a network with external ports
is connected to a router. In this case, it is needed to create a single
scheduler for all the ports connected to the router and matching the
same chassis as the gateway router port.

This method uses the network ID to assign the group name. It also uses
a defined chassis priority (that will match the gateway router port
assignation). The created ``HA_Chassis_Group`` must have a set of
``HA_Chassis`` registers that should have the same chassis name and
priority as the gateway router port ``Gateway_Chassis``.

NOTE: in future developments, the router port ``Gateway_Chassis`` will
be replaced with a ``HA_Chassis_Group`` register.

Related-Bug: #2125553
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: Ia4f685077a8d72bf28f66daf21225d96f57ddef6