Revert "[OVN] Always exclude the local chassis creating ``HA_Chassis_Group``"
This reverts commit 4238ccb45069eb75c1af960d367080e731414a8d.

Reason for revert: this patch is breaking the Ironic CI. The
external can be bound to a gateway chassis and yet this chassis
can be used to build the ``HA_Chassis_Group`` register. This
limitation imposed by the reverted patch doesn't work if the
environment has one single gateway chassis, usually happening
that in CI jobs.

Change-Id: I201d7681c1a1b0c6e3f54c46ba5d4be586c709ff
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>