[OVN] Add 'provider:physical_network' field if there is physnet
In [1], the commit message stated that the field
'provider:physical_network' would be present in the
``Logical_Switch`` if the network has an associated physical network.
If not, the field won't be present.

This field is used in [2], but the code uses ``get`` to retrieve this
field from the dictionary, expecting that this field could be not
present.

[1]https://review.opendev.org/c/openstack/neutron/+/962568
[2]https://github.com/openstack/neutron/blob/1c9bfe52614ba50813dc492bab9f4a24031d7951/neutron/common/ovn/utils.py#L1312

Related-Bug: #2125553
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I9ebf3048cb14fb79e9de612952752c9c56b14420